+++ src/kernel/arch/x86_64/mem/alloc.c
@@ -8,7 +8,7 @@
 #include <mem/alloc.h>
 #include <io/tty.h>
 
-extern uint32_t end;
+extern uint64_t end;
 extern page_directory_t *kernel_directory;
 
 uintptr_t placement_address = (uintptr_t)&end;
@@ -27,7 +27,7 @@ heap_t *create_heap(uintptr_t start, uintptr_t end, uintptr_t max, uint8_t s, ui
 	assert(!end&0xFFF);
 	
 	heap->head = (heap_header_t *)start;
-	heap->head->size = (uint32_t) end - (uint32_t) start;
+	heap->head->size = (uint64_t) end - (uint64_t) start;
 	heap->head->is_free = 1;
 	heap->head->next = NULL;
 	heap->start = start;
@@ -36,12 +36,12 @@ heap_t *create_heap(uintptr_t start, uintptr_t end, uintptr_t max, uint8_t s, ui
 	heap->supervisor = s;
 	heap->readonly = ro;
 	heap_footer_t *foot_ptr = NULL;
-	(foot_ptr = ((heap_footer_t *)((uint32_t)end - sizeof(heap_footer_t))))->size = heap->head->size;
+	(foot_ptr = ((heap_footer_t *)((uint64_t)end - sizeof(heap_footer_t))))->size = heap->head->size;
 	foot_ptr->prev = NULL;
 	return heap;
 }
 
-void *khalloc(uint32_t size, uint8_t align, heap_t *heap)
+void *khalloc(uint64_t size, uint8_t align, heap_t *heap)
 {
 	printf("allocing  0x%x bytes\n", size);
 
@@ -61,13 +61,13 @@ void *khalloc(uint32_t size, uint8_t align, heap_t *heap)
 	}
 	while(!head_ptr->is_free)
 	{
-		head_ptr = (heap_header_t *) prev_contig_block((uintptr_t)((uint32_t)head_ptr + sizeof(heap_header_t)));
+		head_ptr = (heap_header_t *) prev_contig_block((uintptr_t)((uint64_t)head_ptr + sizeof(heap_header_t)));
 		if (!head_ptr)
 			break;
 	}
 	while(head_ptr->size < size + sizeof(heap_header_t) + sizeof(heap_footer_t))
 	{
-		head_ptr = (heap_header_t *) prev_free_block((uintptr_t)((uint32_t)head_ptr + sizeof(heap_header_t)));
+		head_ptr = (heap_header_t *) prev_free_block((uintptr_t)((uint64_t)head_ptr + sizeof(heap_header_t)));
 		if(!head_ptr)
 			break;
 	}
@@ -77,9 +77,9 @@ void *khalloc(uint32_t size, uint8_t align, heap_t *heap)
 		PANIC("Out of heap!!!");	//Something is wrong...
 	}
 
-	heap_footer_t * new_foot = (heap_footer_t *)((uint32_t)head_ptr + head_ptr->size - size - (2 * sizeof(heap_footer_t)));
-	heap_footer_t * foot_ptr = (heap_footer_t *)((uint32_t)head_ptr + head_ptr->size - sizeof(heap_footer_t));
-	heap_header_t * new_head = (heap_header_t *)((uint32_t)new_foot + sizeof(heap_footer_t)); 
+	heap_footer_t * new_foot = (heap_footer_t *)((uint64_t)head_ptr + head_ptr->size - size - (2 * sizeof(heap_footer_t)));
+	heap_footer_t * foot_ptr = (heap_footer_t *)((uint64_t)head_ptr + head_ptr->size - sizeof(heap_footer_t));
+	heap_header_t * new_head = (heap_header_t *)((uint64_t)new_foot + sizeof(heap_footer_t)); 
 
 	new_head->size = size + sizeof(heap_header_t) + sizeof(heap_footer_t);
 	new_head->is_free = 0;
@@ -93,7 +93,7 @@ void *khalloc(uint32_t size, uint8_t align, heap_t *heap)
 	foot_ptr->size = new_head->size;
 	foot_ptr->prev = NULL;
 
-	return (void *)((uint32_t)new_head + sizeof(heap_header_t));
+	return (void *)((uint64_t)new_head + sizeof(heap_header_t));
 }
 
 void khfree(void *p, heap_t *heap)
@@ -103,7 +103,7 @@ void khfree(void *p, heap_t *heap)
 		return;
 	}
 
-	heap_header_t *head_ptr	= (heap_header_t *)((uint32_t) p - sizeof(heap_header_t));
+	heap_header_t *head_ptr	= (heap_header_t *)((uint64_t) p - sizeof(heap_header_t));
 	heap_header_t *next_ptr;
 	heap_header_t *prev_ptr;
 	heap_footer_t *foot_ptr;
@@ -115,10 +115,10 @@ void khfree(void *p, heap_t *heap)
 		if (next_ptr != NULL && next_ptr->is_free)
 		{
 			foot_ptr = (heap_footer_t *)(next_ptr->size
-				+ (uint32_t) next_ptr
+				+ (uint64_t) next_ptr
 				- sizeof(heap_footer_t));
 			foot_ptr->prev = ((heap_footer_t *)
-			(head_ptr->size + (uint32_t) head_ptr - sizeof(heap_footer_t)))->prev;
+			(head_ptr->size + (uint64_t) head_ptr - sizeof(heap_footer_t)))->prev;
 
 			head_ptr->next = next_ptr->next;
 			head_ptr->size += next_ptr->size;
@@ -129,10 +129,10 @@ void khfree(void *p, heap_t *heap)
 		if (prev_ptr != NULL && prev_ptr->is_free)
 		{
 			foot_ptr = (heap_footer_t *)(head_ptr->size 
-				+ (uint32_t) head_ptr 
+				+ (uint64_t) head_ptr 
 				- sizeof(heap_footer_t));
 			foot_ptr->prev = ((heap_footer_t *)
-			(prev_ptr->size +(uint32_t) prev_ptr - sizeof(heap_footer_t)))->prev;
+			(prev_ptr->size +(uint64_t) prev_ptr - sizeof(heap_footer_t)))->prev;
 			
 			prev_ptr->next = head_ptr->next;
 			prev_ptr->size += head_ptr->size;
@@ -143,32 +143,32 @@ void khfree(void *p, heap_t *heap)
 
 uintptr_t next_free_block(uintptr_t u_data)
 {
-	heap_header_t *head_ptr = (heap_header_t*)((uint32_t) u_data - sizeof(heap_header_t));
+	heap_header_t *head_ptr = (heap_header_t*)((uint64_t) u_data - sizeof(heap_header_t));
 	return (head_ptr->is_free) ? head_ptr->next : (uintptr_t) NULL;
 }
 
 uintptr_t next_contig_block(uintptr_t u_data)
 {
-	heap_header_t *head_ptr =  (heap_header_t *)((uint32_t) u_data - sizeof(heap_header_t));
-	if((uint32_t) head_ptr >= kernel_heap->end || (uint32_t) head_ptr + head_ptr->size >= kernel_heap->end) return NULL;
-	return (uintptr_t) (head_ptr->size + (uint32_t) head_ptr);
+	heap_header_t *head_ptr =  (heap_header_t *)((uint64_t) u_data - sizeof(heap_header_t));
+	if((uint64_t) head_ptr >= kernel_heap->end || (uint64_t) head_ptr + head_ptr->size >= kernel_heap->end) return NULL;
+	return (uintptr_t) (head_ptr->size + (uint64_t) head_ptr);
 }
 
 uintptr_t prev_free_block(uintptr_t u_data)
 {
-	heap_header_t *head_ptr = (heap_header_t *)((uint32_t) u_data - sizeof(heap_header_t));
+	heap_header_t *head_ptr = (heap_header_t *)((uint64_t) u_data - sizeof(heap_header_t));
 	heap_footer_t *foot_ptr = (heap_footer_t *)(head_ptr->size + 
-								(uint32_t) head_ptr 
+								(uint64_t) head_ptr 
 								- sizeof(heap_footer_t));
 	return (head_ptr->is_free) ? foot_ptr->prev : (uintptr_t) NULL;
 }
 
 uintptr_t prev_contig_block(uintptr_t u_data)
 {
-	heap_header_t *head_ptr = (heap_header_t*)((uint32_t) u_data - sizeof(heap_header_t));
-	heap_footer_t *prev_foot_ptr = (heap_footer_t *)((uint32_t) head_ptr - sizeof(heap_header_t) - sizeof(heap_footer_t));
-	if((uint32_t) prev_foot_ptr < kernel_heap->start || (uint32_t) head_ptr - prev_foot_ptr->size < kernel_heap->start) return NULL;
-	return (uintptr_t)((uint32_t) head_ptr - prev_foot_ptr->size);
+	heap_header_t *head_ptr = (heap_header_t*)((uint64_t) u_data - sizeof(heap_header_t));
+	heap_footer_t *prev_foot_ptr = (heap_footer_t *)((uint64_t) head_ptr - sizeof(heap_header_t) - sizeof(heap_footer_t));
+	if((uint64_t) prev_foot_ptr < kernel_heap->start || (uint64_t) head_ptr - prev_foot_ptr->size < kernel_heap->start) return NULL;
+	return (uintptr_t)((uint64_t) head_ptr - prev_foot_ptr->size);
 }
